backend/requirements.txt
# FastAPI and Server
fastapi==0.104.1
uvicorn==0.24.0

# Database
psycopg2-binary==2.9.9
python-dotenv==1.0.0

# Security and Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# API Documentation
pydantic==2.5.1
pydantic-settings==2.1.0

# Utilities
python-dateutil==2.8.2
email-validator==2.1.0.post1

# Testing
pytest==7.4.3
httpx==0.25.1

# Development
black==23.10.1
flake8==6.1.0
